Create README.md
I am a young woman who wishes to be in the mining technical industry with any mine that is open to welcome and groom me. I am looking forward to introduce myself and skills to a new working environment that will allow me to grow and improve my working capabilities.